# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает информацию о текущем аутентифицированном пользователе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Получить профиль пользователя",
                "responses": {
                    "200": {
                        "description": "Профиль пользователя",
                        "schema": {
                            "$ref": "#/definitions/response.UserResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не аутентифицирован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет учетную запись текущего аутентифицированного пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Удалить профиль пользователя",
                "responses": {
                    "200": {
                        "description": "Пользователь успешно удален",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не аутентифицирован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/tariffs": {
            "get": {
                "description": "Возвращает детальный расчет стоимости доставки по конкретному тарифу СДЭК",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "delivery"
                ],
                "summary": "Получить расчет тарифа СДЭК",
                "parameters": [
                    {
                        "description": "Параметры для расчета тарифа",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CDEKRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Расчет тарифа",
                        "schema": {
                            "$ref": "#/definitions/response.CDEKTariffCalcResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат входных данных",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tariffslist": {
            "get": {
                "description": "Возвращает список доступных тарифов СДЭК для указанных параметров доставки",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "delivery"
                ],
                "summary": "Получить список тарифов СДЭК",
                "parameters": [
                    {
                        "description": "Параметры для расчета тарифов",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CDEKRequestList"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список тарифов",
                        "schema": {
                            "$ref": "#/definitions/response.CDEKTariffListResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат входных данных",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "request.CDEKRequest": {
            "type": "object",
            "required": [
                "from_address",
                "height",
                "length",
                "tariff_code",
                "to_address",
                "weight",
                "width"
            ],
            "properties": {
                "from_address": {
                    "description": "Адрес отправления",
                    "type": "string",
                    "example": "Москва"
                },
                "height": {
                    "description": "Высота в см",
                    "type": "integer",
                    "example": 15
                },
                "length": {
                    "description": "Длина в см",
                    "type": "integer",
                    "example": 30
                },
                "tariff_code": {
                    "description": "Код тарифа СДЭК",
                    "type": "integer",
                    "example": 139
                },
                "to_address": {
                    "description": "Адрес доставки",
                    "type": "string",
                    "example": "Санкт-Петербург"
                },
                "weight": {
                    "description": "Вес в граммах",
                    "type": "integer",
                    "example": 2500
                },
                "width": {
                    "description": "Ширина в см",
                    "type": "integer",
                    "example": 20
                }
            }
        },
        "request.CDEKRequestList": {
            "type": "object",
            "required": [
                "from_address",
                "height",
                "length",
                "to_address",
                "weight",
                "width"
            ],
            "properties": {
                "from_address": {
                    "description": "Адрес отправления",
                    "type": "string",
                    "example": "Москва"
                },
                "height": {
                    "description": "Высота в см",
                    "type": "integer",
                    "example": 15
                },
                "length": {
                    "description": "Длина в см",
                    "type": "integer",
                    "example": 30
                },
                "to_address": {
                    "description": "Адрес доставки",
                    "type": "string",
                    "example": "Санкт-Петербург"
                },
                "weight": {
                    "description": "Вес в граммах",
                    "type": "integer",
                    "example": 2500
                },
                "width": {
                    "description": "Ширина в см",
                    "type": "integer",
                    "example": 20
                }
            }
        },
...
        "response.CDEKDeliveryDateRange": {
            "type": "object",
            "properties": {
                "max": {
                    "description": "\"2022-02-04\"",
                    "type": "string"
                },
                "min": {
                    "description": "\"2022-02-02\"",
                    "type": "string"
                }
            }
        },
        "response.CDEKErrorMessage": {
            "type": "object",
            "properties": {
                "additional_code": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.CDEKServicePrice": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "discount_percent": {
                    "type": "number"
                },
                "discount_sum": {
                    "type": "number"
                },
                "sum": {
                    "type": "number"
                },
                "total_sum": {
                    "type": "number"
                },
                "vat_rate": {
                    "type": "number"
                },
                "vat_sum": {
                    "type": "number"
                }
            }
        },
        "response.CDEKTariffCalcResponse": {
            "type": "object",
            "properties": {
                "calendar_max": {
                    "type": "integer"
                },
                "calendar_min": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                },
                "delivery_date_range": {
                    "$ref": "#/definitions/response.CDEKDeliveryDateRange"
                },
                "delivery_sum": {
                    "type": "number"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.CDEKErrorMessage"
                    }
                },
                "period_max": {
                    "type": "integer"
                },
                "period_min": {
                    "type": "integer"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.CDEKServicePrice"
                    }
                },
                "total_sum": {
                    "type": "number"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.CDEKWarningMessage"
                    }
                },
                "weight_calc": {
                    "type": "integer"
                }
            }
        },
        "response.CDEKTariffListItem": {
            "type": "object",
            "properties": {
                "calendar_max": {
                    "type": "integer"
                },
                "calendar_min": {
                    "type": "integer"
                },
                "delivery_date_range": {
                    "$ref": "#/definitions/response.CDEKDeliveryDateRange"
                },
                "delivery_mode": {
                    "type": "integer"
                },
                "delivery_sum": {
                    "type": "number"
                },
                "period_max": {
                    "type": "integer"
                },
                "period_min": {
                    "type": "integer"
                },
                "tariff_code": {
                    "type": "integer"
                },
                "tariff_description": {
                    "type": "string"
                },
                "tariff_name": {
                    "type": "string"
                }
            }
        },
        "response.CDEKTariffListResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.CDEKErrorMessage"
                    }
                },
                "tariff_codes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.CDEKTariffListItem"
                    }
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.CDEKWarningMessage"
                    }
                }
            }
        },
        "response.CDEKWarningMessage": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Дата создания",
                    "type": "string",
                    "example": "2024-01-15T10:00:00Z"
                },
                "id": {
                    "description": "UUID пользователя",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "role": {
                    "description": "Роль пользователя",
                    "type": "string",
                    "example": "user"
                },
                "updated_at": {
                    "description": "Дата обновления",
                    "type": "string",
                    "example": "2024-01-15T10:00:00Z"
                },
                "username": {
                    "description": "Имя пользователя",
                    "type": "string",
                    "example": "user123"
                }
            }
        }
    }
}